I had lot of joy with klines development today :).
That's because I have just implemented an "A*" pathfinding algorithm.
I found excellent beginners tutorial at www.gamedev.net.

I wonder if it is needed in other kdegames?
If yes, I think, it is easy to make it more general for other's use.

I failed to understand how the pathfinding is
implemented in the old klines code, there were
no comments, just a pure code with some misterious vars.

So I went to gamedev and found how to do that myself :).

Path is found but not used yet. That's what I'll do next ;).

svn path=/trunk/KDE/kdegames/klines/; revision=616115