Implement one level undo.
Note that due to bug in QTimeLine, after this commit the game becomes unplayable
unless you exchange lines 113 and 114 in qtimeline.cpp which is in
qt-copy/src/corelib/tools/.

I thought about making patch for qt-copy, but this is not a critical one, so
I think it doesn't worth it :-)

Well, in fact I can place a temporary workaround so patching qt-copy won't be needed.
If someone needs this, ping me somewhere :-).

Bug has been reported to TT. Not accepted yet :).

svn path=/trunk/KDE/kdegames/klines/; revision=619586